Code clean up and worked on tests

+++ libfwsi/libfwsi_delegate_values.h
@@ -50,11 +50,11 @@ int libfwsi_delegate_values_free(
      libfwsi_delegate_values_t **delegate_values,
      libcerror_error_t **error );
 
-ssize_t libfwsi_delegate_values_read_data(
-         libfwsi_delegate_values_t *delegate_values,
-         const uint8_t *data,
-         size_t data_size,
-         libcerror_error_t **error );
+int libfwsi_delegate_values_read_data(
+     libfwsi_delegate_values_t *delegate_values,
+     const uint8_t *data,
+     size_t data_size,
+     libcerror_error_t **error );
 
 #if defined( __cplusplus )
 }
