+++ libfwsi/libfwsi_file_entry_values.h
@@ -72,12 +72,12 @@ int libfwsi_file_entry_values_free(
      libfwsi_file_entry_values_t **file_entry_values,
      libcerror_error_t **error );
 
-ssize_t libfwsi_file_entry_values_read_data(
-         libfwsi_file_entry_values_t *file_entry_values,
-         const uint8_t *data,
-         size_t data_size,
-         int ascii_codepage,
-         libcerror_error_t **error );
+int libfwsi_file_entry_values_read_data(
+     libfwsi_file_entry_values_t *file_entry_values,
+     const uint8_t *data,
+     size_t data_size,
+     int ascii_codepage,
+     libcerror_error_t **error );
 
 #if defined( __cplusplus )
 }
