+++ libfwsi/libfwsi_mtp_volume_values.h
@@ -48,11 +48,11 @@ int libfwsi_mtp_volume_values_free(
      libfwsi_mtp_volume_values_t **mtp_volume_values,
      libcerror_error_t **error );
 
-ssize_t libfwsi_mtp_volume_values_read_data(
-         libfwsi_mtp_volume_values_t *mtp_volume_values,
-         const uint8_t *data,
-         size_t data_size,
-         libcerror_error_t **error );
+int libfwsi_mtp_volume_values_read_data(
+     libfwsi_mtp_volume_values_t *mtp_volume_values,
+     const uint8_t *data,
+     size_t data_size,
+     libcerror_error_t **error );
 
 #if defined( __cplusplus )
 }
