+++ libfwsi/libfwsi_network_location_values.h
@@ -72,12 +72,12 @@ int libfwsi_network_location_values_free(
      libfwsi_network_location_values_t **network_location_values,
      libcerror_error_t **error );
 
-ssize_t libfwsi_network_location_values_read_data(
-         libfwsi_network_location_values_t *network_location_values,
-         const uint8_t *data,
-         size_t data_size,
-         int ascii_codepage,
-         libcerror_error_t **error );
+int libfwsi_network_location_values_read_data(
+     libfwsi_network_location_values_t *network_location_values,
+     const uint8_t *data,
+     size_t data_size,
+     int ascii_codepage,
+     libcerror_error_t **error );
 
 #if defined( __cplusplus )
 }
