+++ libfwsi/libfwsi_uri_values.h
@@ -48,12 +48,12 @@ int libfwsi_uri_values_free(
      libfwsi_uri_values_t **uri_values,
      libcerror_error_t **error );
 
-ssize_t libfwsi_uri_values_read_data(
-         libfwsi_uri_values_t *uri_values,
-         const uint8_t *data,
-         size_t data_size,
-         int ascii_codepage,
-         libcerror_error_t **error );
+int libfwsi_uri_values_read_data(
+     libfwsi_uri_values_t *uri_values,
+     const uint8_t *data,
+     size_t data_size,
+     int ascii_codepage,
+     libcerror_error_t **error );
 
 #if defined( __cplusplus )
 }
